# PERF: Sparse Series to scipy COO sparse matrix