--- a/pandas/core/arrays/sparse/scipy_sparse.py
+++ b/pandas/core/arrays/sparse/scipy_sparse.py
@@ -182,7 +182,7 @@ def coo_to_sparse_series(
     Parameters
     ----------
     A : scipy.sparse.coo.coo_matrix
-    dense_index : bool, default FalseSparseDtype
+    dense_index : bool, default False
 
     Returns
     -------