@jreback 2021-09-01
@TLouf lgtm.

@pandas-dev/pandas-core if any comments.